pyproject.toml
[project]
name = "backend-medics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dj-database-url>=2.3.0",
    "aiohappyeyeballs==2.4.3",
    "aiohttp==3.10.10",
    "aiosignal==1.3.1",
    "annotated-types==0.7.0",
    "anyio==4.6.2.post1",
    "asgiref==3.8.1",
    "attrs==24.2.0",
    "certifi==2024.8.30",
    "deprecation==2.1.0",
    "dj-database-url==2.3.0",
    "django==5.1.5",
    "django-cors-headers==4.6.0",
    "django-environ==0.11.2",
    "django-phonenumber-field==8.0.0",
    "django-resized==1.0.3",
    "django-unfold==0.40.0",
    "djangorestframework==3.15.2",
    "frozenlist==1.5.0",
    "gotrue==2.10.0",
    "h11==0.14.0",
    "h2==4.1.0",
    "hpack==4.0.0",
    "httpcore==1.0.6",
    "httpx==0.27.2",
    "hyperframe==6.0.1",
    "idna==3.10",
    "multidict==6.1.0",
    "numpy==2.1.3",
    "packaging==24.1",
    "pandas==2.2.3",
    "phonenumberslite==8.13.49",
    "pillow==11.0.0",
    "postgrest==0.18.0",
    "propcache==0.2.0",
    "psycopg2-binary==2.9.10",
    "pydantic==2.9.2",
    "pydantic-core==2.23.4",
    "python-dateutil==2.9.0.post0",
    "pytz==2024.2",
    "realtime==2.0.6",
    "six==1.16.0",
    "sniffio==1.3.1",
    "sqlparse==0.5.3",
    "storage3==0.9.0",
    "typing-extensions==4.12.2",
    "tzdata==2024.2",
    "websockets==13.1",
    "whitenoise==6.8.2",
    "yarl==1.17.1",
    "minio>=7.2.15",
    "django-minio-storage>=0.5.7",
    "gunicorn>=23.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9.8",
]
